requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMuPDF==1.23.26
Pillow==10.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0